+++ src/Symfony/Component/Notifier/Notification/SmsNotificationInterface.php
@@ -12,7 +12,7 @@
 namespace Symfony\Component\Notifier\Notification;
 
 use Symfony\Component\Notifier\Message\SmsMessage;
-use Symfony\Component\Notifier\Recipient\Recipient;
+use Symfony\Component\Notifier\Recipient\SmsRecipientInterface;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>
@@ -21,5 +21,5 @@ use Symfony\Component\Notifier\Recipient\Recipient;
  */
 interface SmsNotificationInterface
 {
-    public function asSmsMessage(Recipient $recipient, string $transport = null): ?SmsMessage;
+    public function asSmsMessage(SmsRecipientInterface $recipient, string $transport = null): ?SmsMessage;
 }